server/pyproject.toml
[tool.poetry]
name = "maptroid"
version = "0.1.0"
description = ""
authors = ["chriscauley <chris@lablackey.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.2.7"
ImageHash = "^4.2.1"
pytesseract = "^0.3.8"
np = "^1.0.2"
social-auth-app-django = "4.0.0"
social-auth-core = "3.3.3"
django-unrest = "^0.1.2"
psycopg2-binary = "^2.9.2"
uWSGI = "^2.0.20"
django-mailer = "^2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
